requirements.txt
cryptography
certifi
pyOpenSSL